+++ github-page/changelog.md
@@ -47,6 +47,7 @@ redirect_from:
 - exposing settings for "loglevel"
 - fixes Decorator Execution Order [{% include gitissue repo="hasezoey" num=23 %}, {% include gitissue repo="hasezoey" num=24 %}]
 - add support for custom discriminator properties
+- add error if using a self-containing class
 - [IC] "NoParamConstructor" got renamed into "AnyParamConstructor" it now accepts any arguments
 - [IC] Remake data.ts to use Maps hasezoey#3
 - [IC] adding many tests and bumping coverage